IP2Location

Description

This Haskell package provides a fast lookup of country, region, city, latitude, longitude, ZIP code, time zone, ISP, domain name, connection type, IDD code, area code, weather station code, weather station name, mcc, mnc, mobile brand, elevation, and usage type from IP address by using IP2Location database. This package uses a file based database available at IP2Location.com. This database simply contains IP blocks as keys, and other information such as country, region, city, latitude, longitude, ZIP code, time zone, ISP, domain name, connection type, IDD code, area code, weather station code, weather station name, mcc, mnc, mobile brand, elevation, and usage type as values. It supports both IP addresses in IPv4 and IPv6.

IP2Location LITE BIN databases are available for free at http://lite.ip2location.com/

Documentation

data Meta

Contains the BIN database file metadata.

data IP2LocationRecord

Contains geolocation results.

getAPIVersion :: String

The getAPIVersion function returns a string containing the API version.

doInit :: String -> IO Meta

The doInit function returns the Meta record containing metadata from the BIN database file. It takes one argument, of type String, which is the path to the BIN database file.

doQuery :: String -> Meta -> String -> IO IP2LocationRecord

The doQuery function returns an IP2LocationRecord containing geolocation data for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from doInit function (Meta record) & either IPv4 or IPv6 address (String).